requirements.txt
Django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-dotenv==1.0.1
tqdm-4.66.2
faker-24.8.0